Move TestListTile to list_tile_tester.dart
