+++ packages/flutter/test/widgets/autocomplete_test.dart
@@ -10,7 +10,7 @@ import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'editable_text_tester.dart';
-import 'list_tile_test_utils.dart';
+import 'list_tile_tester.dart';
 import 'multi_view_testing.dart';
 import 'widgets_app_tester.dart';
