+++ packages/flutter/test/widgets/reorderable_list_test.dart
@@ -8,7 +8,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'list_tile_test_utils.dart';
+import 'list_tile_tester.dart';
 import 'semantics_tester.dart';
 
 void main() {
