+++ packages/flutter/test/widgets/sliver_main_axis_group_test.dart
@@ -7,7 +7,7 @@ import 'package:flutter/rendering.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import '../rendering/sliver_utils.dart';
-import 'list_tile_test_utils.dart';
+import 'list_tile_tester.dart';
 import 'semantics_tester.dart';
 
 const double VIEWPORT_HEIGHT = 600;
